set_section update

+++ cse351/lab1b/lab1b/aisle_manager.c
@@ -134,7 +134,7 @@ unsigned short get_id(unsigned long *aisle, int index)
 void set_section(unsigned long *aisle, int index, unsigned short new_section)
 {
   // TODO: implement this method
-  *aisle = (*aisle & (~((long)SECTION_MASK << (16 * index)))) | (new_section << (16 * index));
+  *aisle = (*aisle & ~((long)SECTION_MASK << (16 * index))) | ((long)new_section << (16 * index));
 }
 
 /* Given a pointer to an aisle, a section index, and a short representing a new
